// frontend/src/app/shared/generated-api/model/noteSummary.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ChecklistItemView } from './checklistItemView';
import { NoteType } from './noteType';
import { ContentColor } from './contentColor';


export interface NoteSummary { 
    id: string;
    notebookId: string;
    noteType: NoteType;
    title: string;
    preview: string;
    checklistPreview: Array<ChecklistItemView>;
    pinned: boolean;
    archived: boolean;
    color: ContentColor;
    labelIds: Array<string>;
    imageIds: Array<string>;
    deletedAt?: string;
    createdAt: string;
    updatedAt: string;
    version: number;
}